Package refactor, moved custom components to components package
