+++ app/src/main/java/com/itba/runningMate/components/map/Map.kt
@@ -1,4 +1,4 @@
-package com.itba.runningMate.map
+package com.itba.runningMate.components.map
 
 import android.annotation.SuppressLint
 import android.content.Context
