+++ app/src/main/java/com/itba/runningMate/components/map/MapInScrollView.kt
@@ -1,4 +1,4 @@
-package com.itba.runningMate.map
+package com.itba.runningMate.components.map
 
 import android.content.Context
 import android.util.AttributeSet
