--- app/src/main/java/com/itba/runningMate/pastruns/runs/OnRunClickListener.kt
@@ -1,6 +1,0 @@
-package com.itba.runningMate.pastruns.runs
-
-fun interface OnRunClickListener {
-
-    fun onRunClick(id: Long)
-}
