+++ app/src/main/java/com/itba/runningMate/running/fragments/map/RunningMapFragment.kt
@@ -17,7 +17,7 @@ import com.google.android.gms.maps.OnMapReadyCallback
 import com.itba.runningMate.R
 import com.itba.runningMate.di.DependencyContainerLocator
 import com.itba.runningMate.domain.Route
-import com.itba.runningMate.map.Map
+import com.itba.runningMate.components.map.Map
 import com.itba.runningMate.services.location.Tracker
 import com.itba.runningMate.services.location.TrackingService
 import com.itba.runningMate.utils.Constants.MY_LOCATION_ZOOM
