+++ app/src/main/res/layout/fragment_mainpage_running.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-    <com.itba.runningMate.map.MapInViewPager
+    <com.itba.runningMate.components.map.MapInViewPager
         android:id="@+id/map"
         android:name="com.google.android.gms.maps.SupportMapFragment"
         android:layout_width="0dp"
