+++ app/src/main/res/layout/fragment_running_map.xml
@@ -3,7 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-    <com.itba.runningMate.map.MapInViewPager
+    <com.itba.runningMate.components.map.MapInViewPager
         android:id="@+id/map"
         android:name="com.google.android.gms.maps.SupportMapFragment"
         android:layout_width="match_parent"
